* фиксы для дробей (int) на (float) * jbimage добавлен data-template * дефолтное значение селекта для новой почты * Параметр image выбор картинки * Валидация с админки * Стили

+++ branches/cart2/packages/jbuniversal/jbuniversal/framework/classes/cart/jbcart.php
@@ -340,7 +340,7 @@ class JBCart
         if (!empty($items[$key])) {
 
             $item  = $items[$key];
-            $value = (int)$value;
+            $value = (float)$value;
             $value = $value >= 1 ? $value : 1;
 
             $item['quantity'] = $value;
